refator(base dir): clean up the other file names, etc...
