+++ packages/vehicles/src/index.story.js
@@ -7,7 +7,7 @@ import { storiesOf } from "@storybook/react";
 import BaseMap from "@opentripplanner/base-map";
 
 import VehicleLayer from "./VehicleLayer";
-import VehicleAction from "./vehicle-action";
+import VehicleAction from ".";
 
 import "@opentripplanner/base-map/assets/map.css";
